@@ -28,17 +28,17 @@
     "postinstall": "./node_modules/.bin/nx migrate check"
   },
   "dependencies": {
-    "@angular/animations": "8.2.14",
-    "@angular/common": "8.2.14",
-    "@angular/compiler": "8.2.14",
-    "@angular/core": "8.2.14",
+    "@angular/animations": "9.1.7",
+    "@angular/common": "9.1.7",
+    "@angular/compiler": "9.1.7",
+    "@angular/core": "9.1.7",
     "@angular/fire": "^5.4.2",
-    "@angular/forms": "8.2.14",
-    "@angular/platform-browser": "8.2.14",
-    "@angular/platform-browser-dynamic": "8.2.14",
-    "@angular/router": "8.2.14",
+    "@angular/forms": "9.1.7",
+    "@angular/platform-browser": "9.1.7",
+    "@angular/platform-browser-dynamic": "9.1.7",
+    "@angular/router": "9.1.7",
     "@firebase/app": "^0.1.10",
-    "@ngrx/effects": "8.5.1",
+    "@ngrx/effects": "^9.1.2",
     "@ngrx/entity": "8.5.1",
     "@ngrx/router-store": "8.5.1",
     "@ngrx/store": "8.5.1",
@@ -48,24 +48,24 @@
     "core-js": "^2.6.11",
     "firebase": "^7.14.3",
     "firebase-tools": "^7.16.2",
-    "ngx-bootstrap": "^2.0.2",
+    "ngx-bootstrap": "^5.6.1",
     "rxjs": "^6.5.5",
-    "tslib": "^1.13.0",
-    "zone.js": "~0.9.1"
+    "tslib": "^1.10.0",
+    "zone.js": "~0.10.2"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^0.803.26",
-    "@angular-devkit/core": "8.3.19",
-    "@angular-devkit/schematics": "8.3.19",
-    "@angular/cli": "8.3.19",
-    "@angular/compiler-cli": "8.2.14",
-    "@angular/language-service": "8.2.14",
+    "@angular-devkit/build-angular": "~0.901.5",
+    "@angular-devkit/core": "9.1.5",
+    "@angular-devkit/schematics": "9.1.5",
+    "@angular/cli": "9.1.5",
+    "@angular/compiler-cli": "9.1.7",
+    "@angular/language-service": "9.1.7",
     "@nrwl/schematics": "8.7.1",
     "@schematics/angular": "8.3.19",
     "@types/jasmine": "~2.5.53",
     "@types/jasminewd2": "~2.0.2",
-    "@types/node": "~6.0.60",
-    "codelyzer": "^5.2.2",
+    "@types/node": "^12.11.1",
+    "codelyzer": "^5.1.2",
     "jasmine-core": "~2.8.0",
     "jasmine-spec-reporter": "~4.2.1",
     "karma": "^4.4.1",
@@ -78,6 +78,6 @@
     "protractor": "^5.4.4",
     "ts-node": "~4.1.0",
     "tslint": "~5.9.1",
-    "typescript": "3.5.3"
+    "typescript": "3.8.3"
   }
 }
